Splitted code for maintenance purpose
