+++ static/style.css
@@ -8,7 +8,7 @@
   --accent-hover: #79b8ff;
   --danger: #f85149;
   --mono: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace;
-  --sans: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif;
+  --sans: -apple-system, BlinkMacSystemFont, 'Segoe UI', Helvetica, Arial, sans-serif;
 }
 * {
   box-sizing: border-box;
@@ -78,7 +78,9 @@ h1 a:hover {
   font-size: 12px;
   color: var(--muted);
 }
-input, select, button {
+input,
+select,
+button {
   padding: 8px 12px;
   border-radius: 6px;
   border: 1px solid var(--border);
@@ -87,7 +89,8 @@ input, select, button {
   font-size: 14px;
   outline: none;
 }
-input:focus, select:focus {
+input:focus,
+select:focus {
   border-color: var(--accent);
 }
 input::placeholder {
@@ -154,7 +157,8 @@ table {
   width: 100%;
   border-collapse: collapse;
 }
-th, td {
+th,
+td {
   padding: 10px 12px;
   text-align: left;
   border-bottom: 1px solid var(--border);
@@ -169,7 +173,7 @@ th {
   white-space: nowrap;
 }
 th[data-sort] {
-    cursor: pointer;
+  cursor: pointer;
 }
 th:hover {
   color: var(--text);
@@ -321,14 +325,14 @@ tr.highlight td {
   gap: 8px;
   align-items: center;
 }
-.pager-controls input[type="number"] {
+.pager-controls input[type='number'] {
   width: 60px;
   text-align: center;
   -moz-appearance: textfield;
   appearance: textfield;
 }
-.pager-controls input[type="number"]::-webkit-outer-spin-button,
-.pager-controls input[type="number"]::-webkit-inner-spin-button {
+.pager-controls input[type='number']::-webkit-outer-spin-button,
+.pager-controls input[type='number']::-webkit-inner-spin-button {
   -webkit-appearance: none;
   margin: 0;
 }
